sr(api): move training-serializer in devcomp
