+++ api/src/devcomp/infrastructure/serializers/jsonapi/training-serializer.js
