+++ api/tests/devcomp/unit/infrastructure/serializers/jsonapi/training-serializer_test.js
@@ -1,5 +1,5 @@
-import { domainBuilder, expect } from '../../../../test-helper.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/training-serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
+import * as serializer from '../../../../../../src/devcomp/infrastructure/serializers/jsonapi/training-serializer.js';
 
 describe('Unit | Serializer | JSONAPI | training-serializer', function () {
   describe('#serializeForAdmin', function () {
